feat: add CI workflow, Etherscan verification test, and fix clippy warnings
- Add GitHub Actions CI (check, clippy, test, fmt) and Dependabot config
- Add verify_fixtures.rs integration test: cross-verifies known_arb_txs.json
  against Etherscan V2 getLogs API (Swap events per block)
- Update known_arb_txs.json with 7 Etherscan-confirmed cross-DEX arb txs
  (blocks 15537405-15539280, WETH/USDT + WETH/USDC + WETH/DAI pairs)
- Fix simulation_bench.rs: update PoolState fields and detect_v2_arb_opportunity
  signature (adds 4th arg)
- Fix clippy warnings: unnecessary cast, inconsistent digit grouping,
  print_literal, unnecessary lazy eval, manual range contains, useless vec
- Add dev-dependencies for verify_fixtures: reqwest, serde, serde_json, tokio

+++ tests/arbitrage_correctness.rs
@@ -15,9 +15,9 @@ mod arbitrage_correctness {
     fn make_pool(address_str: &str, reserve0: u128, reserve1: u128, fee_bps: u32) -> PoolState {
         // Convert fee_bps to numerator/denominator (standard V2 = 997/1000)
         let (fee_numerator, fee_denominator) = match fee_bps {
-            30 => (997, 1000),                  // 0.3% fee (standard Uniswap V2)
-            25 => (9975, 10000),                // 0.25% fee (custom V2 fork example)
-            _ => (1000 - fee_bps as u32, 1000), // approximate conversion
+            30 => (997, 1000),           // 0.3% fee (standard Uniswap V2)
+            25 => (9975, 10000),         // 0.25% fee (custom V2 fork example)
+            _ => (1000 - fee_bps, 1000), // approximate conversion
         };
 
         PoolState {
